deserializer using fraction::Decimal for amount

+++ src/models/shared.rs
@@ -1,6 +1,8 @@
-// We're using the as real value the one with the most precision available
+use fraction::Decimal;
+
+// Aiming to use the type with the most precision available
 // and consider rounding as a presentation concern.
-pub type Amount = f64;
+pub type Amount = Decimal;
 
 // Let's say u64 are okay values for hypergrowth at this time :)
 pub type OID = u64;
